feat(us-verification): update address verification example
