+++ lob-java-examples/pom.xml
@@ -14,7 +14,7 @@
     <dependency>
       <groupId>com.lob</groupId>
       <artifactId>lob-java</artifactId>
-      <version>5.0.0</version>
+      <version>6.0.0</version>
     </dependency>
 
     <dependency>
